Ensure R_PrintLongString doesn't split words between buffers
CL_ConsolePrint treats text at the end of a message given to it as a complete word for purposes of word-wrapping. When one half of a word in the first buffer from R_PrintLongString is given to it and this can fit on the current line, the first half will be printed on that line. However, when the second half of the word in the second buffer from R_PrintLongString is given to it and it can't fit on the current line, this second half will be printed on the line underneath. By ensuring R_PrintLongString only passes complete words to be printed, the client console is able to word-wrap properly.

This was originally brought to attention by @eternalcodes in PR #827.