Canister ammunition buff changes
Previously it was possible to set enemies on fire with smoke canisters and explosive canisters using the LS-180, it was also impossible to set players on fire with the pulse shotgun and A200 Battle Rifle launcher firemodes using incendiary canisters.

Currently it is still only possible to set players on fire with a direct hit.