Vulkan: Fix edge case during dglow pass with optimized per-stage computations
Collapsed multi-stage shaders during glow pass:

blackimage texture is used on a non-glow bundle and ( ComputeTexCoords(), ComputeColors() ) is skipped.
TESS_ST0 or TESS_RGBA0 are removed on a glow bundle in the next stage when tc and rgb are equal to the non-glow bundle in the previous stage, it will use stale tc and/or rgb data.

Most noticable shader: textures/rooftop/building_ext3 in t2_rogue one particular green building